explicit buckets csv:  * tweaking styles; cleaning up code  * moving functions to the utils file
