+++ cypress/e2e/cloud/buckets.test.ts
@@ -174,75 +174,68 @@ describe('Explicit Buckets', () => {
       })
   })
 
-  it.only('should be able to create an explicit bucket and add schema file during editing', function() {
+  it('should be able to create an explicit bucket and add schema file during editing', function() {
     cy.getByTestID('Create Bucket').click()
-    cy.getByTestID('overlay--container').within(() => {
-      cy.getByTestID('bucket-form-name').type('explicit_bucket')
-      cy.getByTestID('accordion-header').click()
-      cy.getByTestID('explicit-bucket-schema-choice-ID').click()
+    cy.getByTestID('bucket-form-name').type('explicit_bucket')
+    cy.getByTestID('accordion-header').click()
+    cy.getByTestID('explicit-bucket-schema-choice-ID').click()
 
-      cy.getByTestID('bucket-form-submit').click()
-    })
+    cy.getByTestID('bucket-form-submit').click()
+
     cy.getByTestID(`bucket-card explicit_bucket`)
       .should('exist')
       .within(() => {
         cy.getByTestID('bucket-settings').click()
       })
+    cy.getByTestID('accordion-header').click()
 
-    cy.getByTestID('overlay--container').within(() => {
-      cy.getByTestID('accordion-header').click()
+    cy.getByTestID('measurement-schema-add-file-button').click()
+    cy.getByTestID('input-field').type('first schema file')
 
-      cy.getByTestID('measurement-schema-add-file-button').click()
-      cy.getByTestID('input-field').type('first schema file')
-
-      const schemaFile = 'valid.json'
-      const type = 'application/json'
-      const testFile = new File(
-        [
-          `[{"name":"time","type":"timestamp"},
+    const schemaFile = 'valid.json'
+    const type = 'application/json'
+    const testFile = new File(
+      [
+        `[{"name":"time","type":"timestamp"},
         {"name":"fsWrite","type":"field","dataType":"float"} ]`,
-        ],
-        schemaFile,
-        {type}
-      )
+      ],
+      schemaFile,
+      {type}
+    )
 
-      const event = {dataTransfer: {files: [testFile]}, force: true}
-      cy.getByTestID('dndContainer')
-        .trigger('dragover', event)
-        .trigger('drop', event)
+    const event = {dataTransfer: {files: [testFile]}, force: true}
+    cy.getByTestID('dndContainer')
+      .trigger('dragover', event)
+      .trigger('drop', event)
 
-      cy.getByTestID('bucket-form-submit').click()
-    })
+    cy.getByTestID('bucket-form-submit').click()
 
     cy.getByTestID(`bucket-card explicit_bucket`)
       .should('exist')
       .within(() => {
         cy.getByTestID('bucket-settings').click()
       })
+    cy.getByTestID('accordion-header').click()
 
-    cy.getByTestID('overlay--container').within(() => {
-      cy.getByTestID('accordion-header').click()
+    cy.getByTestID('measurement-schema-readOnly-panel-0')
+      .should('exist')
+      .within(() => {
+        cy.getByTestID('measurement-schema-name-0')
+          .contains('first schem...')
+          .should('exist')
 
-      cy.getByTestID('measurement-schema-readOnly-panel-0')
-        .should('exist')
-        .within(() => {
-          cy.getByTestID('measurement-schema-name-0')
-            .contains('first schem...')
-            .should('exist')
+        cy.getByTestID('measurement-schema-download-button').click()
+        cy.readFile(`cypress/downloads/first_schema_file.json`)
+          .should('exist')
+          .then(fileContent => {
+            expect(fileContent[0].name).to.be.equal('time')
+            expect(fileContent[0].type).to.be.equal('timestamp')
 
-          cy.getByTestID('measurement-schema-download-button').click()
-          cy.readFile(`cypress/downloads/first_schema_file.json`)
-            .should('exist')
-            .then(fileContent => {
-              expect(fileContent[0].name).to.be.equal('time')
-              expect(fileContent[0].type).to.be.equal('timestamp')
-
-              expect(fileContent[1].name).to.be.equal('fsWrite')
-              expect(fileContent[1].type).to.be.equal('field')
-              expect(fileContent[1].dataType).to.be.equal('float')
-            })
-        })
-    })
+            expect(fileContent[1].name).to.be.equal('fsWrite')
+            expect(fileContent[1].type).to.be.equal('field')
+            expect(fileContent[1].dataType).to.be.equal('float')
+          })
+      })
   })
 
   it('should be able to create an explicit bucket and update the existing schema file during editing', function() {
