+++ src/buckets/components/createBucketForm/MeasurementSchema.scss
@@ -32,8 +32,20 @@
     }
   }
 }
-
 .schema-section {
+  .option {
+    margin-bottom: 6px !important;
+  }
+}
+
+.measurement-section {
+  .option:nth-child(3) {
+    margin-left: 10px;
+  }
+}
+
+.schema-section,
+.measurement-section {
   &.measurement {
     margin-top: 15px;
   }
@@ -83,10 +95,6 @@
   .value-text {
     color: $cf-white;
     font-weight: 500;
-  }
-
-  .option {
-    margin-bottom: 6px !important;
   }
 
   .measurement-schema-panel-container {
