+++ src/buckets/components/createBucketForm/MeasurementSchemaSection.test.tsx
@@ -5,6 +5,7 @@ import {
   TOO_LONG_ERROR,
   areNewSchemasValid,
   areSchemaUpdatesValid,
+  csvToArray,
 } from './MeasurementSchemaUtils'
 import {toCsvString} from './MeasurementSchemaSection'
 
@@ -283,5 +284,42 @@ describe('test csv conversion function (object->csv file)', () => {
     expect(lines[9]).toEqual('one,field,string')
 
     expect(lines.length).toEqual(10)
+  })
+})
+describe('test csv to array function (parsing)', () => {
+  it('should parse the csv correctly', () => {
+    const contents = `foo,bar,baz
+    hello,how,are
+    you,today,really`
+
+    const parsed = csvToArray(contents)
+
+    expect(parsed?.length).toEqual(2)
+    expect(parsed[0]['foo']).toEqual('hello')
+    expect(parsed[0]['bar']).toEqual('how')
+    expect(parsed[0]['baz']).toEqual('are')
+    expect(parsed[1]['foo']).toEqual('you')
+    expect(parsed[1]['bar']).toEqual('today')
+    expect(parsed[1]['baz']).toEqual('really')
+  })
+  it('should parse the csv correctly, with a missing third element', () => {
+    const contents = `foo,bar,baz
+    hello,how,are
+    you,today,`
+
+    const parsed = csvToArray(contents)
+
+    expect(parsed?.length).toEqual(2)
+    expect(parsed[0]['foo']).toEqual('hello')
+    expect(parsed[0]['bar']).toEqual('how')
+    expect(parsed[0]['baz']).toEqual('are')
+    expect(parsed[1]['foo']).toEqual('you')
+    expect(parsed[1]['bar']).toEqual('today')
+
+    const secondKeys = Object.keys(parsed[1])
+    expect(secondKeys.length).toEqual(2)
+    expect(secondKeys).toContain('foo')
+    expect(secondKeys).toContain('bar')
+    expect(secondKeys).not.toContain('baz')
   })
 })
